Move data classes to their own module and reorganise ml_models module
