+++ backend/server/ml_models/__init__.py
@@ -1,4 +1,5 @@
 """ML model classes and associated data classes"""
 
+from .base_ml_model import BaseMLModel
 from .all_model import AllModel
 from .ensemble_model import EnsembleModel
