+++ backend/server/ml_models/all_model/__init__.py
@@ -1,1 +1,1 @@
-from .all_model import AllModel, AllModelData
+from .all_model import AllModel
