--- backend/server/ml_models/betting_model/__init__.py
@@ -1,1 +1,0 @@
-from .betting_model import BettingModelData
