+++ backend/server/ml_models/ensemble_model/ensemble_model.py
@@ -8,9 +8,9 @@ from sklearn.compose import ColumnTransformer
 from sklearn.ensemble import BaggingRegressor
 from xgboost import XGBRegressor
 
-from server.ml_models.ml_model import MLModel
 from server.data_config import TEAM_NAMES, ROUND_TYPES, VENUES, CATEGORY_COLS
 from server.ml_models.sklearn import CorrelationSelector
+from .. import BaseMLModel
 
 SEED = 42
 np.random.seed(SEED)
@@ -47,7 +47,7 @@ PIPELINE = make_pipeline(
 ).set_params(**BEST_PARAMS)
 
 
-class EnsembleModel(MLModel):
+class EnsembleModel(BaseMLModel):
     """Model for averaging predictions of an ensemble of models"""
 
     def __init__(
