--- backend/server/ml_models/match_model/__init__.py
@@ -1,1 +1,0 @@
-from .match_model import MatchModelData
