--- backend/server/ml_models/player_model/__init__.py
@@ -1,1 +1,0 @@
-from .player_model import PlayerModelData
